requirements.txt
langchain~=0.3.25
langchain-core~=0.3.61
langchain-openai~=0.2.14
langchain-community~=0.3.24
python-dotenv~=1.1.0
dashscope~=1.23.3